02-analytics-dashboard/requirements.txt
streamlit>=1.29.0
plotly>=5.17.0
pandas>=2.1.0
supabase>=2.7.0
python-dotenv>=1.0.0